+++ groups/urls.py
@@ -2,7 +2,7 @@
 
 from django.urls import path
 
-from . import views
+from . import feeds, views
 
 #: The URL namespace of the groups app.
 app_name = 'groups'
@@ -11,6 +11,8 @@ app_name = 'groups'
 urlpatterns = [
     path('', views.all_groups, name='all_groups'),
     path('<int:g_id>/', views.group, name='group'),
+    path('<int:g_id>.atom', feeds.GroupAtom(), name='group.atom'),
+    path('<int:g_id>.rss', feeds.GroupRSS(), name='group.rss'),
 ]
 
 __all__ = ['app_name', 'urlpatterns']