+++ reader/migrations/0002_series_created.py
@@ -1,0 +1,33 @@
+from django.db import migrations, models
+
+
+def populate_dates(apps, schema_editor):
+    series = apps.get_model('reader', 'Series')
+    series._meta.get_field('modified').auto_now = False
+    for s in series.objects.prefetch_related('chapters').all():
+        ch = s.chapters.only('uploaded').earliest('uploaded')
+        s.created = ch.uploaded
+        s.save()
+    series._meta.get_field('modified').auto_now = True
+
+
+class Migration(migrations.Migration):
+    dependencies = [('reader', '0001_squashed')]
+
+    operations = [
+        migrations.AddField(
+            model_name='series',
+            name='created',
+            field=models.DateTimeField(
+                auto_now_add=False, db_index=True, null=True
+            )
+        ),
+        migrations.RunPython(populate_dates, migrations.RunPython.noop),
+        migrations.AlterField(
+            model_name='series',
+            name='created',
+            field=models.DateTimeField(
+                auto_now_add=True, db_index=True, null=False
+            )
+        ),
+    ]