+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+v0.6.4
+^^^^^^
+
+* Added creation date to series
+* Added library, releases, bookmarks RSS feeds
+
 v0.6.3
 ^^^^^^